🔄 `parsers`: Extract `BaseStdoutParser` to dedicated module
Move `BaseStdoutParser` from `outputs/parsers/base.py` into a new
`outputs/parsers/stdout.py`, so `parsers/base.py` contains only the fully generic
`BaseOutputFileParser` and has zero `qe_tools.*` imports or QE references. This is a
prerequisite for carving out a generic package: `BaseStdoutParser` is
QE-specific (it matches the `Program <NAME> v.X.Y` banner and calls
`convert_qe_time_to_sec`), so it does not belong in a code-agnostic base module.

Pure relocation — no behavior changes. Call sites in `parsers/pw.py` (`PwStdoutParser`
subclass) and `outputs/dos.py` (direct use for the `dos.x` stdout) are updated to import
from the new location.
