+++ src/qe_tools/outputs/parsers/pw.py
@@ -6,7 +6,8 @@ from xml.etree import ElementTree
 from xmlschema import XMLSchema
 
 from qe_tools.outputs.parsers import schemas
-from qe_tools.outputs.parsers.base import BaseOutputFileParser, BaseStdoutParser
+from qe_tools.outputs.parsers.base import BaseOutputFileParser
+from qe_tools.outputs.parsers.stdout import BaseStdoutParser
 
 
 class PwXMLParser(BaseOutputFileParser):
